My Second project
this project is not done yet. I need to complete the user's dashboard that I'm currently working on. Also, there's the admin's dashboard that I haven't started on yet.
it's an E-commerce project. We sell electronics supplies. Its database is also in the file. It contains 3 languages: PHP, HTML, JavaScript.